✨ Unique Attacks on Dice Roll! 🚚 Restructured Directory 🍻
I made an oopsie and did not atomically define this commit (did not commit after the restruc).

Directory is organized to anticipate a library of unique attacks.

Unique attacks on dice roll! The dice system will now call on a dice object and draw an index from the dice array to pick a unique attack face. There is only one unique attack in the library, but attacks are only made through the dice system for all units.